feat(catalog): change product price type to float and update proto files
Update product price field from int32 to float32 to support decimal pricing.
Remove duplicate pb files and update protoc version in generated files.
Implement gRPC server methods for product operations.

+++ catalog/models.go
@@ -1,14 +1,14 @@
 package catalog
 
 type Product struct {
-	ID          string `json:"id"`
-	Name        string `json:"name"`
-	Description string `json:"description"`
-	Price       int    `json:"price"`
+	ID          string  `json:"id"`
+	Name        string  `json:"name"`
+	Description string  `json:"description"`
+	Price       float32 `json:"price"`
 }
 
 type ProductDocument struct {
-	Name        string `json:"name"`
-	Description string `json:"description"`
-	Price       int    `json:"price"`
+	Name        string  `json:"name"`
+	Description string  `json:"description"`
+	Price       float32 `json:"price"`
 }
